backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pypdf2==3.0.1
mistralai==1.2.5
pydantic==2.9.2
python-dotenv==1.0.1
